requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-firebase==1.2
requests==2.6.0
spotipy==2.3.3
Werkzeug==0.10.4
nltk==3.0.2
